pyproject.toml
 [tool.poetry]
 name = "recruiter-market-brief"
 version = "0.1.0"
 description = "Recruiter Market Brief Phase 1 backend and data pipeline"
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
 packages = [
   { include = "backend", from = "." }
 ]

 [tool.poetry.dependencies]
 python = "^3.11"
 python-dotenv = "1.0.0"
 pydantic = "2.5.0"
 structlog = "24.1.0"
 supabase = "2.3.0"
 "psycopg[binary]" = "3.1.16"
 apify-client = "1.6.0"
 requests = "2.31.0"
 tenacity = "8.2.3"
 pandas = "2.1.4"
 numpy = "1.26.3"
 jinja2 = "3.1.3"
 markdown = "3.5.1"

 [tool.poetry.group.dev.dependencies]
 pytest = "7.4.3"
 pytest-cov = "4.1.0"
 pytest-mock = "3.12.0"
 faker = "22.0.0"
 ruff = "0.1.9"
 black = "23.12.1"
 mypy = "1.8.0"
 pre-commit = "3.6.0"

 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"

 [tool.black]
 line-length = 88
 target-version = ["py311"]

 [tool.ruff]
 line-length = 88
 target-version = "py311"
 src = ["backend", "tests"]

 [tool.mypy]
 python_version = "3.11"
 files = ["backend", "tests"]
 ignore_missing_imports = true
 strict = false

